pyproject.toml
[tool.poetry]
name = "passport"
version = "0.4.0"
description = "Packing list management tool"
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.10"
python-dotenv = "^1.0.0"
django-filter = "^23.2"
django-bootstrap-v5 = "^1.0.11"
django-cachalot = "^2.5.2"
django-ajax-selects = "^2.2.0"
django-tables2 = "^2.5.1"
tablib = {extras = ["ods", "xlsx"], version = "^3.3.0"}
django-widget-tweaks = "^1.4.12"
pillow = "^9.4.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.4.0"
uvicorn = "^0.22.0"
django-cors-headers = "^4.1.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.8.1"
django-stubs = "^1.14.0"

[tool.poetry.group.testing.dependencies]
faker = "^18.4.0"
coveralls = "^3.3.1"
model-bakery = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source = ['.']
omit = ['*/tests/*', '*/migrations/*', 'passport/*']

[tool.coverage.report]
skip_covered = true

[tool.ruff]
line-length = 119
exclude = ['*/migrations/*']
select = ['E', 'F', 'UP', 'NPY', 'SIM', 'ICN', 'C4', 'A', 'B', 'N', 'I', 'Q', 'DJ',
    'W', 'PL', 'RET', 'RUF', 'PERF', 'PIE', 'ARG', 'PTH', 'PD', 'PGH', 'TRY']
ignore = ['TRY003', 'ARG002', 'RUF012']

[tool.ruff.flake8-quotes]
inline-quotes = 'single'
multiline-quotes = 'double'
